refactor v1 tests
